Finished first pass of full path for copying and renaming files. This will work for the most common situations. Need to fix:
There is a case where one card of a two card set has a gap in files and those
files are on the second card.  If the card with the gap is loaded first, then
the second card does not load because that last file of the first card is newer
than the files on the second card. We need to add code to find these gaps and
load the files that fill the gaps into the correct directory.