Added preview panes and code to populate them by selecting lines in the card summary list box. The code works for both picture files and video files (tested so far on jpg, nef and mov). The video file preview is just the first frame of the video, which is fine for the purpose.
To do: more testing and cleanup, figure out a way to pass
context to the list box selection callback to get rid of
the global variables in there now.

Later: add time shifting functionality. Work out how to
read in a field archive (backups saved on a trip to portable
storage)